Rename the module from accounts to oscar_accounts
The module name `accounts` is far to generic and can easily conflict.
This change is obviously not backwards compatbile
