+++ src/oscar_accounts/checkout/gateway.py
@@ -3,7 +3,7 @@ from django.utils.translation import ugettext_lazy as _
 from oscar.apps.payment.exceptions import UnableToTakePayment
 from oscar.core.loading import get_model
 
-from accounts import facade, exceptions, core, codes
+from oscar_accounts import facade, exceptions, core, codes
 
 Account = get_model('accounts', 'Account')
 Transfer = get_model('accounts', 'Transfer')
