+++ src/oscar_accounts/config.py
@@ -2,7 +2,7 @@ from django.apps import AppConfig
 from django.utils.translation import ugettext_lazy as _
 
 
-class AccountsConfig(AppConfig):
+class OscarAccountsConfig(AppConfig):
     label = 'accounts'
-    name = 'accounts'
+    name = 'oscar_accounts'
     verbose_name = _('Accounts')
