+++ src/oscar_accounts/management/commands/oscar_accounts_init.py
@@ -1,6 +1,6 @@
 from django.core.management.base import BaseCommand
 
-from accounts.setup import create_default_accounts
+from oscar_accounts.setup import create_default_accounts
 
 
 class Command(BaseCommand):
