+++ src/oscar_accounts/templates/accounts/dashboard/partials/account_detail.html
