+++ src/oscar_accounts/views.py
@@ -1,7 +1,7 @@
 from django.views import generic
 
-from accounts.forms import AccountForm
-from accounts import security
+from oscar_accounts.forms import AccountForm
+from oscar_accounts import security
 
 
 class AccountBalanceView(generic.FormView):
