+++ tests/unit/facade_tests.py
@@ -6,9 +6,9 @@ from django.test import TestCase, TransactionTestCase
 from oscar.test.factories import UserFactory
 import mock
 
-from accounts import facade, exceptions
-from accounts.models import Account, Transfer, Transaction
-from accounts.test_factories import AccountFactory
+from oscar_accounts import facade, exceptions
+from oscar_accounts.models import Account, Transfer, Transaction
+from oscar_accounts.test_factories import AccountFactory
 
 
 class TestReversingATransfer(TestCase):
@@ -107,7 +107,7 @@ class TestErrorHandling(TransactionTestCase):
         source = AccountFactory(credit_limit=None)
         destination = AccountFactory()
         with mock.patch(
-            'accounts.abstract_models.PostingManager._wrap') as mock_method:
+            'oscar_accounts.abstract_models.PostingManager._wrap') as mock_method:
             mock_method.side_effect = RuntimeError()
             try:
                 facade.transfer(source, destination, D('100'), user)
@@ -128,7 +128,7 @@ class TestErrorHandling(TransactionTestCase):
         source = AccountFactory(credit_limit=None)
         destination = AccountFactory()
         with mock.patch(
-            'accounts.abstract_models.PostingManager._wrap') as mock_method:
+            'oscar_accounts.abstract_models.PostingManager._wrap') as mock_method:
             mock_method.side_effect = RuntimeError()
             with self.assertRaises(exceptions.AccountException):
                 facade.transfer(source, destination, D('100'), user)
