+++ tests/unit/transfer_tests.py
@@ -5,9 +5,9 @@ from django.test import TestCase
 from django.utils import timezone
 from oscar.test.factories import UserFactory
 
-from accounts import exceptions
-from accounts.models import Account, Transfer
-from accounts.test_factories import AccountFactory
+from oscar_accounts import exceptions
+from oscar_accounts.models import Account, Transfer
+from oscar_accounts.test_factories import AccountFactory
 
 
 class TestASuccessfulTransfer(TestCase):
